Title & Q4 BMI
When I ran the equation the first time, I entered the wrong BMI.

The answer using the equation or the prediction are nearly identical, except that the prediction comes with the confidence interval built in.

I also updated the top of the page to show our group and the correct due date.